feat: don't spawn new scratchpad on different monitor
Currently the scratchpad is stored on a per-monitor basis. So if it does not
exist on each monitor it is spawned on each one. This commit extends the
togglescratch function to search for the scratchpad across all monitors and
re-attach it to the current one. The window size is adjusted to the current
monitor dimensions.

Signed-off-by: Martin Homuth <martin.homuth@emlix.com>